port the player screen to webpack 4
