--- .ebextensions/compile.config
@@ -1,15 +1,0 @@
-#commands:
-#  01-install-nodejs-npm:
-#    command: "yum install -y --enablerepo=epel nodejs npm"
-
-container_commands:
-  01-npm-install:
-    env:
-      HOME: "/root"
-      PATH: "/usr/bin:/usr/lib/node_modules:/sbin:/bin:/usr/sbin:/opt/aws/bin:/opt/elasticbeanstalk/node-install/node-v8.4.0-linux-x64/bin"
-    command: "npm install"
-  02-compilecss-minifyjs:
-    env:
-      HOME: "/root"
-      PATH: "/usr/bin:/usr/lib/node_modules:/sbin:/bin:/usr/sbin:/opt/aws/bin:/opt/elasticbeanstalk/node-install/node-v8.4.0-linux-x64/bin"
-    command: "npm run gulp"
