--- .elasticbeanstalk/config.yml
@@ -1,21 +1,0 @@
-branch-defaults:
-  master:
-    environment: prod
-    group_suffix: null
-environment-defaults:
-  prod:
-    branch: master
-    repository: cityofmist
-global:
-  application_name: cityofmist
-  branch: master
-  default_ec2_keyname: null
-  default_platform: Node.js
-  default_region: us-east-1
-  instance_profile: null
-  platform_name: null
-  platform_version: null
-  profile: eb-cli
-  repository: cityofmist
-  sc: git
-  workspace_type: Application
