+++ index.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html>
+	<head>
+		<meta charset="utf-8" /> 
+		<title>City of Mist</title>
+	</head>
+	<body>
+        <script src="https://use.fontawesome.com/c9e2e20754.js"></script>
+		
+		<div id="root"></div>
+	</body>
+</html>
