+++ package.json
@@ -6,12 +6,21 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "webpack",
+    "start": "webpack-dev-server",
     "gulp": "gulp"
   },
   "author": "",
   "license": "ISC",
-  "dependencies": {},
+  "dependencies": {
+    "faye": "^1.2.4",
+    "jquery": "^3.3.1",
+    "mithril": "^1.1.6"
+  },
   "devDependencies": {
+    "clean-webpack-plugin": "^0.1.19",
+    "css-loader": "^0.28.11",
+    "favicons-webpack-plugin": "0.0.7",
+    "file-loader": "^1.1.11",
     "gulp": "^3.9.1",
     "gulp-concat": "^2.6.1",
     "gulp-gzip": "^1.4.2",
@@ -19,8 +28,12 @@
     "gulp-sourcemaps": "^2.6.4",
     "gulp-uglify": "latest",
     "gulp-util": "latest",
+    "html-webpack-plugin": "^3.0.6",
     "pump": "^1.0.3",
+    "style-loader": "^0.20.3",
     "uglify-es": "^3.3.10",
-    "webpack": "^4.1.1"
+    "webpack": "^4.1.1",
+    "webpack-cli": "^2.0.12",
+    "webpack-dev-server": "^3.1.1"
   }
 }
