+++ src/Deck.js
@@ -1,4 +1,14 @@
-var Deck = {
+import m from 'mithril';
+import Name from './Name';
+import TabSwitcher from './TabSwitcher';
+import Theme from './Theme';
+import Roller from './Roller';
+import Statuses from './Statuses';
+import Moves from './Moves';
+import MapScreen from './MapScreen';
+import { firstName, toggleLocked } from './common';
+
+export default {
 	view: function(vnode) {
 		return [
 			m(TabSwitcher, {
@@ -23,7 +33,7 @@ var Deck = {
 			m("#main", [
 				m(Name, {char: vnode.attrs.char}),
 				m(Theme, {themes: vnode.attrs.char.themes}),
-				m(Roller, {rolls: vnode.attrs.rolls, who: firstName(vnode.attrs.char.name), room: myRoom, char: vnode.attrs.char }),
+				m(Roller, { who: firstName(vnode.attrs.char.name), room: vnode.attrs.room, char: vnode.attrs.char }),
 				m(Statuses, {statuses: vnode.attrs.char.statuses}),
 				m("button[class=unlock]", { onclick: () => {toggleLocked()}}, "Lock/unlock themes")
 			]),
