+++ src/images/minus-lt.gif
