--- static/index.html
@@ -1,20 +1,0 @@
-<!doctype html>
-<html>
-	<head>
-		<meta charset="utf-8" /> 
-		<title>City of Mist</title>
-		<link href="index.css" rel="stylesheet" />
-		<link href="images/city-tiny.png" rel="icon" />
-		<script src="https://code.jquery.com/jquery-3.2.1.min.js"></script>
-        
-	</head>
-	<body>
-        <script src="https://unpkg.com/mithril/mithril.js"></script>
-        <script src="/faye/client.js"></script>
-        
-        <script src="index.js"></script>
-        <script src="https://use.fontawesome.com/c9e2e20754.js"></script>
-		
-		<div id="root"></div>
-	</body>
-</html>
